Got to build on ghcjs frontend - moved regex into example and set for ghc builds only - moved internal library into main library, was causing build issue   - cabal sublibs have some problem with ghcjs - small code fixes
