+++ test/Data/ValidationSpec.hs
@@ -1,7 +1,7 @@
-{-# LANGUAGE 
+{-# LANGUAGE
     FlexibleInstances
   , MultiParamTypeClasses
-  , TemplateHaskellQuotes 
+  , TemplateHaskellQuotes
   , TypeSynonymInstances
 #-}
 
@@ -36,11 +36,11 @@ spec = parallel $ do
 
     it "If applied to two invalid proofs, it concatenates the errors." $ do
       let
-        vi1 = Invalid ["Failure 1"] 
+        vi1 = Invalid ["Failure 1"]
           (fromList [([mkName "Field1"], ["Field Failure 1"])])
-        vi2 = Invalid ["Failure 2"] 
+        vi2 = Invalid ["Failure 2"]
           (fromList [([mkName "Field1"], ["Field Failure 1.1"]), ([mkName "Field2"], ["Field Failure 2"])])
-        vi3 = Invalid ["Failure 1", "Failure 2"] 
+        vi3 = Invalid ["Failure 1", "Failure 2"]
           (fromList [([mkName "Field1"], ["Field Failure 1", "Field Failure 1.1"]), ([mkName "Field2"], ["Field Failure 2"])])
       (vi1 :: Proof String [Int]) <> vi2 `shouldBe` vi3
 
@@ -53,7 +53,7 @@ spec = parallel $ do
       fmap show (vi :: Proof String Int) `shouldBe` (vi :: Proof String String)
 
   describe "Applicative (Proof f)" $ do
-    let 
+    let
       vap = Valid show :: Proof String (Int -> String)
       iap = Invalid ["Applicative Failure"] mempty
       vp = Valid 1 :: Proof String Int
@@ -63,11 +63,11 @@ spec = parallel $ do
       pure 1 `shouldBe` (Valid 1 :: Proof String Int)
 
     it "If applied to two invalid proofs, it concatenates the errors." $ do
-      (iap :: Proof String (Int -> String)) <*> ip 
+      (iap :: Proof String (Int -> String)) <*> ip
         `shouldBe` Invalid ["Applicative Failure", "Failure"] (fromList [([mkName "Field1"], ["Field Failure"])])
 
     it "If applied to an invalid and valid proof, it results in the invalid proof." $ do
-      (iap :: Proof String (Int -> String)) <*> vp 
+      (iap :: Proof String (Int -> String)) <*> vp
         `shouldBe` (iap :: Proof String String)
 
     it "If applied to a valid and invalid proof, it results in the invalid proof." $ do
@@ -101,34 +101,34 @@ spec = parallel $ do
 
   describe "Functor ValueCtx" $ do
     it "transforms a global context." $ do
-      fmap show (Global 1 :: ValueCtx Int) `shouldBe` Global "1" 
+      fmap show (Global 1 :: ValueCtx Int) `shouldBe` Global "1"
 
     it "transforms a field context while preserving the field name." $ do
-      fmap show (Field (mkName "Field1") 1 :: ValueCtx Int) `shouldBe` Field (mkName "Field1") "1" 
+      fmap show (Field (mkName "Field1") 1 :: ValueCtx Int) `shouldBe` Field (mkName "Field1") "1"
 
   describe "getValue" $ do
     it "Retrieves the value from a global context." $ do
       getValue (Global "1") `shouldBe` "1"
-      
+
     it "Retrieves the value from a field context." $ do
       getValue (Field (mkName "Field1") "1") `shouldBe` "1"
 
   describe "setValue" $ do
     it "Creates a global context based on the given context." $ do
       setValue (Global "1") (1 :: Int) `shouldBe` Global 1
-      
+
     it "Retrieves the value from a field context." $ do
       setValue (Field (mkName "Field1") "1") (1 :: Int) `shouldBe` (Field (mkName "Field1") 1)
 
   describe "isRequired" $ do
     it "Adds a failure to the context if the value is Nothing." $ do
-      let 
+      let
         v = Global (Nothing :: Maybe Bool)
         r = isRequired "Failed" v
       r `shouldBe` refute v "Failed"
 
     it "Should add no failure if the value is Just." $ do
-      let 
+      let
         v = Global $ Just True
         v2 = Global True
         r = isRequired "Failed" v
@@ -136,13 +136,13 @@ spec = parallel $ do
 
   describe "isLeft" $ do
     it "Adds a failure to the context if the value is Right." $ do
-      let 
+      let
         v = Global (Right True :: Either Bool Bool)
         r = isLeft "Failed" v
       r `shouldBe` refute v "Failed"
 
     it "Should add no failure if the value is Left." $ do
-      let 
+      let
         v = Global $ (Left True :: Either Bool Bool)
         v2 = Global True
         r = isLeft "Failed" v
@@ -150,13 +150,13 @@ spec = parallel $ do
 
   describe "isRight" $ do
     it "Adds a failure to the context if the value is Left." $ do
-      let 
+      let
         v = Global (Left True :: Either Bool Bool)
         r = isRight "Failed" v
       r `shouldBe` refute v "Failed"
 
     it "Should add no failure if the value is Right." $ do
-      let 
+      let
         v = Global $ (Right True :: Either Bool Bool)
         v2 = Global True
         r = isRight "Failed" v
@@ -164,109 +164,109 @@ spec = parallel $ do
 
   describe "isNull" $ do
     it "Adds a failure to the context if the value is not null." $ do
-      let 
+      let
         v = Global "not null"
         r = isNull "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is null." $ do
-      let 
+      let
         v = Global ""
         r = isNull "Failed" v
       r `shouldBe` pure v
 
   describe "isNotNull" $ do
     it "Adds a failure to the context if the value is null." $ do
-      let 
+      let
         v = Global ""
         r = isNotNull "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is not null." $ do
-      let 
+      let
         v = Global "not null"
         r = isNotNull "Failed" v
       r `shouldBe` pure v
 
   describe "minLength" $ do
     it "Adds a failure to the context if the value has a length less than the given value." $ do
-      let 
+      let
         v = Global "TW"
         r = minLength 3 "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value has a length equal to the given value." $ do
-      let 
+      let
         v = Global "THR"
         r = minLength 3 "Failed" v
       r `shouldBe` pure v
 
     it "Should add no failure if the value has a length greater than the given value." $ do
-      let 
+      let
         v = Global "THREE"
         r = minLength 3 "Failed" v
       r `shouldBe` pure v
 
   describe "maxLength" $ do
     it "Adds a failure to the context if the value has a length greater than the given value." $ do
-      let 
+      let
         v = Global "FOUR"
         r = maxLength 3 "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value has a length equal to the given value." $ do
-      let 
+      let
         v = Global "THR"
         r = maxLength 3 "Failed" v
       r `shouldBe` pure v
 
     it "Should add no failure if the value has a length less than the given value." $ do
-      let 
+      let
         v = Global "TW"
         r = maxLength 3 "Failed" v
       r `shouldBe` pure v
 
   describe "isLength" $ do
     it "Adds a failure to the context if the value has a length greater than the given value." $ do
-      let 
+      let
         v = Global "FOUR"
         r = isLength 3 "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value has a length equal to the given value." $ do
-      let 
+      let
         v = Global "THR"
         r = isLength 3 "Failed" v
       r `shouldBe` pure v
 
     it "Adds a failure to the context if the value has a length less than the given value." $ do
-      let 
+      let
         v = Global "TW"
         r = isLength 3 "Failed" v
       r `shouldBe` dispute v "Failed"
 
   describe "isEqual" $ do
     it "Adds a failure to the context if the value is not equal to the given value." $ do
-      let 
+      let
         v = Global "1"
         r = isEqual "2" "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is equal to the given value." $ do
-      let 
+      let
         v = Global "1"
         r = isEqual "1" "Failed" v
       r `shouldBe` pure v
 
   describe "isNotEqual" $ do
     it "Adds a failure to the context if the value is equal to the given value." $ do
-      let 
+      let
         v = Global "1"
         r = isNotEqual "1" "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is not equal to the given value." $ do
-      let 
+      let
         v = Global "1"
         r = isNotEqual "2" "Failed" v
       r `shouldBe` pure v
@@ -276,21 +276,21 @@ spec = parallel $ do
       one = 1 :: Int
       two = 2 :: Int
       three = 3 :: Int
-      
+
     it "Adds a failure to the context if the value is greater than the given value." $ do
-      let 
+      let
         v = Global three
         r = isLessThan two "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Adds a failure to the context if the value is equal to the given value." $ do
-      let 
+      let
         v = Global two
         r = isLessThan two "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is less than the given value." $ do
-      let 
+      let
         v = Global one
         r = isLessThan two "Failed" v
       r `shouldBe` pure v
@@ -300,21 +300,21 @@ spec = parallel $ do
       one = 1 :: Int
       two = 2 :: Int
       three = 3 :: Int
-      
+
     it "Adds a failure to the context if the value is greater than the given value." $ do
-      let 
+      let
         v = Global three
         r = isLessThanOrEqual two "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value is equal to the given value." $ do
-      let 
+      let
         v = Global two
         r = isLessThanOrEqual two "Failed" v
       r `shouldBe` pure v
 
     it "Should add no failure if the value is less than the given value." $ do
-      let 
+      let
         v = Global one
         r = isLessThanOrEqual two "Failed" v
       r `shouldBe` pure v
@@ -324,21 +324,21 @@ spec = parallel $ do
       one = 1 :: Int
       two = 2 :: Int
       three = 3 :: Int
-      
+
     it "Should add no failure if the value is greater than the given value." $ do
-      let 
+      let
         v = Global three
         r = isGreaterThan two "Failed" v
       r `shouldBe` pure v
 
     it "Adds a failure to the context if the value is equal to the given value." $ do
-      let 
+      let
         v = Global two
         r = isGreaterThan two "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Adds a failure to the context if the value is less than the given value." $ do
-      let 
+      let
         v = Global one
         r = isGreaterThan two "Failed" v
       r `shouldBe` dispute v "Failed"
@@ -348,114 +348,101 @@ spec = parallel $ do
       one = 1 :: Int
       two = 2 :: Int
       three = 3 :: Int
-      
+
     it "Should add no failure if the value is greater than the given value." $ do
-      let 
+      let
         v = Global three
         r = isGreaterThanOrEqual two "Failed" v
       r `shouldBe` pure v
 
     it "Should add no failure if the value is equal to the given value." $ do
-      let 
+      let
         v = Global two
         r = isGreaterThanOrEqual two "Failed" v
       r `shouldBe` pure v
 
     it "Adds a failure to the context if the value is less than the given value." $ do
-      let 
+      let
         v = Global one
         r = isGreaterThanOrEqual two "Failed" v
       r `shouldBe` dispute v "Failed"
 
   describe "hasElem" $ do
     it "Should add no failure if the value contains the given element." $ do
-      let 
+      let
         v = Global "test@example.com"
         r = hasElem '@' "Failed" v
       r `shouldBe` pure v
 
     it "Adds a failure to the context if the value does not contain the given element." $ do
-      let 
+      let
         v = Global "test.example.com"
         r = hasElem '@' "Failed" v
       r `shouldBe` dispute v "Failed"
 
   describe "doesNotHaveElem" $ do
     it "Adds a failure to the context if the value contains the given element." $ do
-      let 
+      let
         v = Global "test@example site.com"
         r = doesNotHaveElem ' ' "Failed" v
       r `shouldBe` dispute v "Failed"
 
     it "Should add no failure if the value does not contain the given element." $ do
-      let 
+      let
         v = Global "test@example-site.com"
         r = doesNotHaveElem ' ' "Failed" v
       r `shouldBe` pure v
 
   describe "ifAny" $ do
     it "Adds a failure to the context if no element in the list passes the check." $ do
-      let 
+      let
         v = Global ["foo", "bar", "baz", "bat"]
         r = ifAny (\e -> bool (Just (e ++ " does not equal buz")) Nothing $ e == "buz") v
       r `shouldBe` disputeMany v ["foo does not equal buz","bar does not equal buz","baz does not equal buz","bat does not equal buz"]
 
     it "Should add no failure if any element in the list passed the check." $ do
-      let 
+      let
         v = Global ["foo", "bar", "baz", "bat"]
         r = ifAny (\e -> bool (Just (e ++ " does not equal foo")) Nothing $ e == "foo") v
       r `shouldBe` pure v
 
   describe "ifAll" $ do
     it "Adds a failure to the context if any element in the list does not pass the check." $ do
-      let 
+      let
         v = Global ["foo", "bar", "baz", "bat"]
         r = ifAll (\e -> bool (Just (e ++ " does not have an 'a'")) Nothing $ elem 'a' e) v
       r `shouldBe` disputeMany v ["foo does not have an 'a'"]
 
     it "Should add no failure if every element in the list passed the check." $ do
-      let 
+      let
         v = Global ["bar", "baz", "bat"]
         r = ifAll (\e -> bool (Just (e ++ " does not have an 'a'")) Nothing $ elem 'a' e) v
       r `shouldBe` pure v
 
   describe "ifEach" $ do
     it "Adds a failure to the context if any element in the list does not pass the check." $ do
-      let 
+      let
         v = Global ["1", "2", "3", "bat"]
         r = ifEach (\e -> readEither e :: Either String Int) v
       r `shouldBe` refute v "Prelude.read: no parse"
 
     it "Should add no failure if every element in the list passed the check." $ do
-      let 
+      let
         v = Global ["1", "2", "3"]
         r = ifEach (\e -> readEither e :: Either String Int) v
         v2 = Global [1, 2, 3]
       r `shouldBe` pure v2
 
-  describe "matchesRegex" $ do
-    it "Adds a failure to the context if the string does not match the pattern." $ do
-      let 
-        v = Global "asdf"
-        r = matchesRegex "\\`[0-9]+\\'" "Failure" v
-      r `shouldBe` dispute v "Failure"
-
-    it "Should add no failure if the string matches the pattern." $ do
-      let 
-        v = Global "1234"
-        r = matchesRegex "\\`[0-9]+\\'" "Failure" v
-      r `shouldBe` pure v
-
   describe "isMatch" $ do
     it "Adds a failure to the context if value does not match the given value." $ do
-      let 
+      let
         v = Global "asdf"
         vm = pure "fdsa"
         r = isMatch "Failure" vm v
       r `shouldBe` dispute v "Failure"
 
     it "Should add no failure if the value matches the given value." $ do
-      let 
+      let
         v = Global "asdf"
         vm = pure "asdf"
         r = isMatch "Failure" vm v
@@ -480,8 +467,8 @@ data ContactVM
   } deriving (Show, Eq)
 
 instance Validatable String ContactVM Contact where
-  validation c = 
-    let 
+  validation c =
+    let
       vp = withField 'phoneNumber (phoneNumber c) $ \v ->
         isNotNull "Phone Number cannot be empty." v
     in pure Contact <*> vp
